revert(autonomous): remove worktree node_modules shim
The shim (e6fba763) executed via 'sudo -u <owner> bash -c', a shape
sudoers rejects BY DESIGN (#2650 root-RCE surface), so it never
succeeded on multi-user prod — and its fail-soft path recorded a
duplicate frontend_node_modules_shim_failed milestone on every advance
(5.2k rows across 5 workflows in 4 days; a parked wait-phase workflow
polling every ~13s contributed 3.7k).

It also hardcoded this repo's own frontend/ layout, so as a product
feature it never generalized: any other project shape was already a
no-op. Dependency setup moves to the agent (lockfile-detected install
guidance + failure-signature translation, tracked in the follow-up) —
zero config, no new sudoers surface.

Removes: _ensure_frontend_node_modules_shim(_impl), the script builder
+ cache-dir constants, both ensure_worktree call sites, and
GitHubOps._run_as_account (its only caller). The 5 shim tests go with
the code; the cleanup tests (worktree_path retention, #2505) remain in
the renamed test_git_workspace_cleanup.py.

Refs open-ace/open-ace#2694 (replaces PR #2695's wrapper approach)
