--- tests/autonomous/test_git_workspace_node_modules_shim.py
@@ -1,209 +1,0 @@
-"""Worktree node_modules shim (#23).
-
-The autonomous agent runs frontend tests (vitest) in its worktree, but the
-worktree has no ``frontend/node_modules`` (gitignored), so the agent falls back
-to the project owner's main-clone node_modules and hits EACCES on
-``node_modules/.vite-temp`` (vite writes its bundled config there). The shim
-injects a worktree-local node_modules: symlinks each package from the main clone
-but creates the writable cache dirs (``.vite-temp``/``.vite``/``.vitest``/
-``.cache``) as REAL worktree dirs (covered by the existing per-worktree ACL grant).
-"""
-
-from __future__ import annotations
-
-import os
-import subprocess
-
-from app.modules.workspace.autonomous.git_workspace import (
-    _FRONTEND_NODE_MODULES_CACHE_DIRS,
-    _build_node_modules_shim_script,
-)
-
-
-def _make_main_clone_with_node_modules(main_root: str) -> str:
-    """Create a fake main clone with frontend/package.json + node_modules.
-
-    Returns the path to ``frontend/node_modules``. Includes a package dir, a
-    dotfile, and a pre-existing cache dir (stand-in for dwu-owned cache the
-    agent must NOT write through).
-    """
-    main_fe = os.path.join(main_root, "frontend")
-    main_nm = os.path.join(main_fe, "node_modules")
-    os.makedirs(os.path.join(main_nm, "react"))
-    with open(os.path.join(main_nm, "react", "index.js"), "w") as f:
-        f.write("x")
-    os.makedirs(os.path.join(main_nm, ".bin"))
-    with open(os.path.join(main_nm, ".package-lock.json"), "w") as f:
-        f.write("{}")
-    # a cache dir that already exists in the main clone (must become REAL in wt,
-    # not a symlink to this dwu-owned one)
-    os.makedirs(os.path.join(main_nm, ".vite-temp"))
-    with open(os.path.join(main_fe, "package.json"), "w") as f:
-        f.write("{}")
-    return main_nm
-
-
-def test_shim_symlinks_packages_and_makes_cache_dirs_real(tmp_path):
-    """Packages are symlinked from the main clone; cache dirs are REAL worktree
-    dirs (writable by the agent via the existing ACL grant), not symlinks."""
-    main_nm = _make_main_clone_with_node_modules(str(tmp_path / "main"))
-    wt_fe = tmp_path / "wt" / "frontend"
-    wt_fe.mkdir(parents=True)
-
-    script = _build_node_modules_shim_script(str(wt_fe), main_nm)
-    res = subprocess.run(["bash", "-c", script], capture_output=True, text=True)
-    assert res.returncode == 0, res.stderr
-
-    wt_nm = wt_fe / "node_modules"
-    # packages → symlinks to the main clone
-    assert (wt_nm / "react").is_symlink()
-    assert os.path.realpath(wt_nm / "react") == os.path.join(main_nm, "react")
-    assert (wt_nm / ".bin").is_symlink()
-    assert (wt_nm / ".package-lock.json").is_symlink()
-    # cache dirs → REAL dirs (not symlinks), writable
-    for cache in _FRONTEND_NODE_MODULES_CACHE_DIRS:
-        cache_path = wt_nm / cache
-        assert cache_path.is_dir(), f"{cache} should be a real dir"
-        assert not cache_path.is_symlink(), f"{cache} must NOT be a symlink (EACCES)"
-        assert os.access(cache_path, os.W_OK), f"{cache} should be writable"
-
-
-def test_shim_idempotent_skips_when_node_modules_populated(tmp_path):
-    """If the worktree already has a populated node_modules (shimmed before, or
-    the agent ran npm install), the shim is a no-op — it must not clobber it."""
-    main_nm = _make_main_clone_with_node_modules(str(tmp_path / "main"))
-    wt_fe = tmp_path / "wt" / "frontend"
-    wt_fe.mkdir(parents=True)
-    # pre-populate with something the agent created
-    (wt_fe / "node_modules").mkdir()
-    (wt_fe / "node_modules" / "agent-installed").mkdir()
-
-    script = _build_node_modules_shim_script(str(wt_fe), main_nm)
-    res = subprocess.run(["bash", "-c", script], capture_output=True, text=True)
-    assert res.returncode == 0, res.stderr
-
-    # the agent's content is preserved; nothing symlinked in
-    assert (wt_fe / "node_modules" / "agent-installed").is_dir()
-    assert not (wt_fe / "node_modules" / "react").exists()
-
-
-def test_shim_atomic_no_partial_state_on_missing_main(tmp_path):
-    """If the main clone has no node_modules, the shim is a no-op and leaves no
-    partial node_modules behind (clean fallback, not a broken half-state)."""
-    main_fe = tmp_path / "main" / "frontend"
-    main_fe.mkdir(parents=True)
-    (main_fe / "package.json").write_text("{}")  # frontend but NO node_modules
-    wt_fe = tmp_path / "wt" / "frontend"
-    wt_fe.mkdir(parents=True)
-
-    script = _build_node_modules_shim_script(str(wt_fe), str(main_fe / "node_modules"))
-    res = subprocess.run(["bash", "-c", script], capture_output=True, text=True)
-    # no main node_modules → nothing to link; the guard / empty-glob handles it
-    # cleanly, exit 0, and no partial node_modules with broken symlinks.
-    assert res.returncode == 0, res.stderr
-    wt_nm = wt_fe / "node_modules"
-    if wt_nm.exists():
-        # if created, it must not contain broken symlinks
-        for entry in wt_nm.iterdir():
-            assert entry.exists() or entry.is_symlink(), f"broken entry {entry}"
-
-
-def test_ensure_shim_fail_soft_swallows_error_and_milestones():
-    """The public wrapper runs on EVERY worktree-strategy workflow's critical
-    path, so it must NEVER raise: an impl error is swallowed and recorded as a
-    ``frontend_node_modules_shim_failed`` milestone, leaving the worktree usable.
-    """
-    from unittest.mock import MagicMock
-
-    from app.modules.workspace.autonomous.git_workspace import GitWorkspaceService
-
-    svc = GitWorkspaceService(MagicMock())
-    wf = {"current_phase": "development"}
-
-    def _boom(*args, **kwargs):
-        raise RuntimeError("boom")
-
-    svc._ensure_frontend_node_modules_shim_impl = _boom  # type: ignore[assignment]
-
-    # must not raise
-    svc._ensure_frontend_node_modules_shim("/tmp/wt", MagicMock(name="gh"), wf)
-
-    # the failure was recorded as a milestone (auditable), not raised
-    svc._orch._create_milestone.assert_called_once()
-    kwargs = svc._orch._create_milestone.call_args.kwargs
-    assert kwargs["milestone_type"] == "frontend_node_modules_shim_failed"
-    assert kwargs["status"] == "failed"
-    assert "boom" in kwargs["error_message"]
-
-
-def test_cleanup_failed_keeps_worktree_path_for_recreate(tmp_path):
-    """A terminally-failed workflow's worktree DIR is removed but ``worktree_path``
-    is KEPT so :meth:`ensure_worktree` recreates the worktree on retry/reset (#23).
-
-    The cleanup docstring already promises "_ensure_worktree recreates it on the
-    next cycle if retried/resumed", but clearing ``worktree_path`` here defeated
-    that: ``ensure_worktree``'s empty-path guard then falls back to the main
-    clone instead of recreating — so a reset failed workflow re-ran in the main
-    clone and hit EACCES on ``node_modules/.vite-temp`` (c88afdc0/83ffb529/
-    ee678c63). The completed/merged path still clears it (merged = done).
-    """
-    from unittest.mock import MagicMock, patch
-
-    from app.modules.workspace.autonomous.git_workspace import GitWorkspaceService
-
-    orch = MagicMock()
-    orch._workflow_id = "wid-failed-keep"
-    orch.workflow = {
-        "branch_name": "auto-dev/x",
-        "worktree_path": str(tmp_path / "wt"),
-        "project_path": str(tmp_path / "main"),
-        "user_id": None,
-    }
-    updates: list[dict] = []
-    orch._update_workflow.side_effect = lambda d: updates.append(d)
-
-    fake_gh = MagicMock()
-    fake_gh.has_uncommitted_changes.return_value = False  # clean → dir removed
-
-    with patch("app.modules.workspace.autonomous.orchestrator.GitHubOps", return_value=fake_gh):
-        svc = GitWorkspaceService(orch)
-        ok = svc.cleanup_worktree_and_branch(reason="failed", remove_worktree=True)
-
-    assert ok is True
-    fake_gh.remove_worktree.assert_called_once()
-    # worktree_path must NOT be cleared (no update setting it to "")
-    cleared = [u for u in updates if u.get("worktree_path") == ""]
-    assert cleared == [], f"failed-cleanup must keep worktree_path for recreate, got {updates}"
-
-
-def test_cleanup_completed_clears_worktree_path(tmp_path):
-    """The completed/merged path still clears ``worktree_path`` (the merged
-    workflow is done; ensure_worktree's empty-path guard correctly returns the
-    main clone for any post-merge probe)."""
-    from unittest.mock import MagicMock, patch
-
-    from app.modules.workspace.autonomous.git_workspace import GitWorkspaceService
-
-    orch = MagicMock()
-    orch._workflow_id = "wid-completed-clear"
-    orch.workflow = {
-        "branch_name": "auto-dev/y",
-        "worktree_path": str(tmp_path / "wt"),
-        "project_path": str(tmp_path / "main"),
-        "user_id": None,
-    }
-    updates: list[dict] = []
-    orch._update_workflow.side_effect = lambda d: updates.append(d)
-
-    fake_gh = MagicMock()  # completed path: no dirty guard, reclaimed directly
-
-    with patch("app.modules.workspace.autonomous.orchestrator.GitHubOps", return_value=fake_gh):
-        svc = GitWorkspaceService(orch)
-        ok = svc.cleanup_worktree_and_branch(
-            reason="completed", remove_worktree=True, remove_branch=True
-        )
-
-    assert ok is True
-    # worktree_path IS cleared for completed
-    cleared = [u for u in updates if u.get("worktree_path") == ""]
-    assert cleared, f"completed-cleanup must clear worktree_path, got {updates}"
